code update

+++ app/Http/Controllers/Grades/GradeController.php
@@ -6,6 +6,7 @@ use toastr;
 use App\Models\Grade;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreGrades;
+use App\Models\Classroom;
 use App\Http\Controllers\Controller;
 
 class GradeController extends Controller
@@ -29,6 +30,11 @@ class GradeController extends Controller
      */
     public function store(StoreGrades $request)
     {
+        // if (Grade::where('Name->ar', $request->Name)->orWhere('Name->en', $request->Name_en)->exists()){
+        //     return redirect()->back()->withErrors(trans('Grades_trans.exists'));
+        // }
+
+
         try {
             $validated = $request->validated();
             $Grades = new Grade();
@@ -45,7 +51,9 @@ class GradeController extends Controller
             return redirect()->route('Grades.index');
             toastr()->success(trans('messages.success'));
             return redirect()->route('Grades.index');
-        } catch (\Exception $e) {
+        }
+        
+        catch (\Exception $e) {
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
     }
@@ -82,11 +90,18 @@ class GradeController extends Controller
         */
        public function destroy(Request $request)
        {
-               $Grades = Grade::findOrFail($request->id)->delete();
-               toastr()->error(trans('messages.Delete'));
-               return redirect()->route('Grades.index');
-           
-     
-     
+        $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
+
+        if($MyClass_id->count() == 0){
+  
+            $Grades = Grade::findOrFail($request->id)->delete();
+            toastr()->error(trans('messages.Delete'));
+            return redirect()->route('Grades.index');
+        }
+  
+        else{
+            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+            return redirect()->route('Grades.index');
+        }
        }
     }
